# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel CRUD API Documentation",
        "description": "API for managing users, posts, and comments",
        "contact": {
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Получить список всех постов",
                "operationId": "getPostsList",
                "responses": {
                    "200": {
                        "description": "Список постов успешно получен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PostResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Создать новый пост",
                "operationId": "createPost",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StorePostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Пост успешно создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    }
                }
            }
        },
        "/api/posts/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Получить пост по ID",
                "operationId": "getPostById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Пост успешно найден",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Пост не найден"
                    }
                }
            },
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Обновить существующий пост",
                "operationId": "updatePost",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdatePostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Пост успешно обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/PostResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Запрещено"
                    },
                    "404": {
                        "description": "Пост не найден"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Удалить пост",
                "operationId": "deletePost",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Пост успешно удален"
                    },
                    "403": {
                        "description": "Запрещено"
                    },
                    "404": {
                        "description": "Пост не найден"
                    }
                }
            }
        },
        "/api/users/{id}/active-posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Получить список активных постов пользователя по его ID",
                "operationId": "getActivePostsByUser",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список активных постов успешно получен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PostResource"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Пользователь не найден"
                    }
                }
            }
        },
        "/api/my-posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Получить список постов текущего аутентифицированного пользователя",
                "operationId": "getMyPosts",
                "responses": {
                    "200": {
                        "description": "Список постов успешно получен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/PostResource"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Не авторизован"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "PostResource": {
                "title": "Post Resource",
                "description": "Post resource model",
                "properties": {
                    "id": {
                        "description": "ID of the post",
                        "type": "integer",
                        "format": "int64"
                    },
                    "body": {
                        "description": "The body of the post",
                        "type": "string"
                    },
                    "status": {
                        "description": "The status of the post",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Creation timestamp",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Last update timestamp",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "StorePostRequest": {
                "title": "Store Post Request",
                "description": "Request body for creating a post",
                "required": [
                    "body"
                ],
                "properties": {
                    "body": {
                        "description": "The body of the post",
                        "type": "string"
                    },
                    "status": {
                        "description": "The status of the post",
                        "type": "string"
                    }
                },
                "type": "object"
            },
            "UpdatePostRequest": {
                "title": "Update Post Request",
                "description": "Request body for updating a post",
                "properties": {
                    "body": {
                        "description": "The body of the post",
                        "type": "string"
                    },
                    "status": {
                        "description": "The status of the post",
                        "type": "string"
                    }
                },
                "type": "object"
            }
        }
    },
...
}